pyproject.toml
[project]
name = "carvo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "django-environ>=0.12.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "cloudinary>=1.41.0",
    "django-cloudinary-storage>=0.3.0",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework>=3.16.1",
    "sendgrid>=6.12.5",
    "django-sendgrid-v5>=1.3.1",
    "celery>=5.6.2",
    "redis>=7.1.0",
    "django-celery-results>=2.6.0",
    "django-redis>=6.0.0",
]

[dependency-groups]
dev = [
    "faker>=33.1.0",
]
